better error display
