+++ Misc/NEWS.d/next/C API/2021-03-18-12-44-33.bpo-43541.ICigzd.rst
@@ -1,0 +1,2 @@
+Fix a ``PyEval_EvalCodeEx()`` regression: fix reference counting on
+builtins. Patch by Victor Stinner.
