+++ Misc/NEWS.d/next/Library/2021-03-16-16-05-02.bpo-43521.mRT6fh.rst
@@ -1,0 +1,1 @@
+``ast.unparse`` can now render NaNs and empty sets.
