+++ Tools/peg_generator/pegen/grammar_parser.py
@@ -2,10 +2,7 @@
 # @generated by pegen from ./Tools/peg_generator/pegen/metagrammar.gram
 
 import ast
-import sys
-import tokenize
-
-from typing import Any, Optional
+from typing import Optional
 
 from pegen.parser import memoize, memoize_left_rec, logger, Parser
 from ast import literal_eval
