+++ Tools/peg_generator/pegen/python_generator.py
@@ -27,10 +27,7 @@ MODULE_PREFIX = """\
 # @generated by pegen from {filename}
 
 import ast
-import sys
-import tokenize
-
-from typing import Any, Optional
+from typing import Optional
 
 from pegen.parser import memoize, memoize_left_rec, logger, Parser
